arch: relocate N3-specific proxyStore to providers/rdfjs/n3 keeping quad-store purely abstract
